Clarify intent of providing a configuration as a dependency
The behavior of declaring a `Configuration` as a dependency using
the dependency handler DSL was confusing. Using a direct notation,
this is equivalent to calling `extendsFrom`: it doesn't effectively
_add_ a new dependency to the graph.

However, since we added a `Provider` taking dependency notation,
it was possible to supply a `Configuration` in the provider, in
which case it would be resolved and the _files_ would be added
as dependencies.

This behavior is not intended and certainly confusing. This commit
deprecates the DSL which allows declaring configurations as dependencies,
and makes the build fail if a `Configuration` is used in a `Provider`.

Future Gradle releases may include _different_ DSL methods to support:

- adding a _list_ of dependencies lazily (using a `ListProvider`)
- adding a _file collection_ lazily

Closes #13977